removed megaparse + fix requirements
