--- core/quivr_core/processor/megaparse/__init__.py
@@ -1,3 +1,0 @@
-from .Converter import MegaParse
-
-__all__ = ["MegaParse"]
