+++ core/requirements-dev.lock
@@ -26,6 +26,7 @@ anthropic==0.36.1
 anyio==4.6.2.post1
     # via anthropic
     # via httpx
+    # via openai
 asttokens==2.4.1
     # via stack-data
 attrs==24.2.0
@@ -42,6 +43,8 @@ charset-normalizer==3.4.0
     # via requests
 click==8.1.7
     # via black
+cohere==5.11.1
+    # via langchain-cohere
 comm==0.2.2
     # via ipykernel
 dataclasses-json==0.6.7
@@ -56,10 +59,15 @@ distlib==0.3.9
     # via virtualenv
 distro==1.9.0
     # via anthropic
+    # via openai
 execnet==2.1.1
     # via pytest-xdist
 executing==2.1.0
     # via stack-data
+faiss-cpu==1.9.0
+    # via quivr-core
+fastavro==1.9.7
+    # via cohere
 filelock==3.16.1
     # via huggingface-hub
     # via transformers
@@ -80,10 +88,13 @@ httpcore==1.0.6
     # via httpx
 httpx==0.27.2
     # via anthropic
+    # via cohere
     # via langgraph-sdk
     # via langsmith
+    # via openai
     # via quivr-core
 httpx-sse==0.4.0
+    # via cohere
     # via langgraph-sdk
 huggingface-hub==0.25.2
     # via tokenizers
@@ -104,6 +115,7 @@ jedi==0.19.1
     # via ipython
 jiter==0.6.1
     # via anthropic
+    # via openai
 jsonpatch==1.33
     # via langchain-core
 jsonpointer==3.0.0
@@ -118,15 +130,25 @@ langchain==0.2.16
     # via quivr-core
 langchain-anthropic==0.1.23
     # via quivr-core
+langchain-cohere==0.2.4
+    # via quivr-core
 langchain-community==0.2.17
+    # via langchain-experimental
     # via quivr-core
 langchain-core==0.2.41
     # via langchain
     # via langchain-anthropic
+    # via langchain-cohere
     # via langchain-community
+    # via langchain-experimental
+    # via langchain-openai
     # via langchain-text-splitters
     # via langgraph
     # via langgraph-checkpoint
+    # via quivr-core
+langchain-experimental==0.0.65
+    # via langchain-cohere
+langchain-openai==0.1.25
     # via quivr-core
 langchain-text-splitters==0.2.4
     # via langchain
@@ -166,20 +188,29 @@ nest-asyncio==1.6.0
 nodeenv==1.9.1
     # via pre-commit
 numpy==1.26.4
+    # via faiss-cpu
     # via langchain
     # via langchain-community
+    # via pandas
     # via transformers
+openai==1.52.0
+    # via langchain-openai
 orjson==3.10.7
     # via langgraph-sdk
     # via langsmith
 packaging==24.1
     # via black
+    # via faiss-cpu
     # via huggingface-hub
     # via ipykernel
     # via langchain-core
     # via marshmallow
     # via pytest
     # via transformers
+pandas==2.2.3
+    # via langchain-cohere
+parameterized==0.9.0
+    # via cohere
 parso==0.8.4
     # via jedi
 pathspec==0.12.1
@@ -211,11 +242,14 @@ pycodestyle==2.12.1
     # via flake8
 pydantic==2.9.2
     # via anthropic
+    # via cohere
     # via langchain
     # via langchain-core
     # via langsmith
+    # via openai
     # via quivr-core
 pydantic-core==2.23.4
+    # via cohere
     # via pydantic
 pyflakes==3.2.0
     # via flake8
@@ -229,8 +263,11 @@ pytest==8.3.3
 pytest-asyncio==0.24.0
 pytest-benchmark==4.0.0
 pytest-xdist==3.6.1
-python-dateutil==2.9.0.post0
+python-dateutil==2.8.2
     # via jupyter-client
+    # via pandas
+pytz==2024.2
+    # via pandas
 pyyaml==6.0.2
     # via huggingface-hub
     # via langchain
@@ -245,6 +282,7 @@ regex==2024.9.11
     # via tiktoken
     # via transformers
 requests==2.32.3
+    # via cohere
     # via huggingface-hub
     # via langchain
     # via langchain-community
@@ -268,25 +306,31 @@ sniffio==1.3.1
     # via anthropic
     # via anyio
     # via httpx
+    # via openai
 sqlalchemy==2.0.36
     # via langchain
     # via langchain-community
 stack-data==0.6.3
     # via ipython
+tabulate==0.9.0
+    # via langchain-cohere
 tenacity==8.5.0
     # via langchain
     # via langchain-community
     # via langchain-core
 tiktoken==0.8.0
+    # via langchain-openai
     # via quivr-core
 tokenizers==0.20.1
     # via anthropic
+    # via cohere
     # via transformers
 tornado==6.4.1
     # via ipykernel
     # via jupyter-client
 tqdm==4.66.5
     # via huggingface-hub
+    # via openai
     # via transformers
 traitlets==5.14.3
     # via comm
@@ -299,20 +343,27 @@ transformers==4.45.2
     # via quivr-core
 types-pyyaml==6.0.12.20240917
     # via quivr-core
+types-requests==2.32.0.20241016
+    # via cohere
 typing-extensions==4.12.2
     # via anthropic
+    # via cohere
     # via huggingface-hub
     # via ipython
     # via langchain-core
     # via mypy
+    # via openai
     # via pydantic
     # via pydantic-core
     # via sqlalchemy
     # via typing-inspect
 typing-inspect==0.9.0
     # via dataclasses-json
+tzdata==2024.2
+    # via pandas
 urllib3==2.2.3
     # via requests
+    # via types-requests
 virtualenv==20.26.6
     # via pre-commit
 wcwidth==0.2.13
