+++ core/tests/processor/pdf/test_unstructured_pdf_processor.py
@@ -3,13 +3,17 @@ from uuid import uuid4
 
 import pytest
 from quivr_core.files.file import FileExtension, QuivrFile
-from quivr_core.processor.implementations.default import UnstructuredPDFProcessor
+
+unstructured = pytest.importorskip("unstructured")
 
 all_but_pdf = list(filter(lambda ext: ext != ".pdf", list(FileExtension)))
 
 
+@pytest.mark.unstructured
 @pytest.mark.asyncio
 async def test_unstructured_pdf_processor():
+    from quivr_core.processor.implementations.default import UnstructuredPDFProcessor
+
     p = Path("./tests/processor/pdf/sample.pdf")
     f = QuivrFile(
         id=uuid4(),
@@ -24,9 +28,12 @@ async def test_unstructured_pdf_processor():
     assert len(result) > 0
 
 
+@pytest.mark.unstructured
 @pytest.mark.parametrize("ext", all_but_pdf)
 @pytest.mark.asyncio
 async def test_unstructured_pdf_processor_fail(ext):
+    from quivr_core.processor.implementations.default import UnstructuredPDFProcessor
+
     p = Path("./tests/processor/pdf/sample.pdf")
     f = QuivrFile(
         id=uuid4(),
