+++ core/tests/test_config.py
@@ -4,18 +4,15 @@ from quivr_core.config import LLMEndpointConfig, RetrievalConfig
 def test_default_llm_config():
     config = LLMEndpointConfig()
 
-    assert (
-        config.model_dump()
-        == LLMEndpointConfig(
-            model="gpt-3.5-turbo-0125",
-            llm_base_url=None,
-            llm_api_key=None,
-            max_input_tokens=2000,
-            max_output_tokens=2000,
-            temperature=0.7,
-            streaming=True,
-        ).model_dump()
-    )
+    assert config.model_dump(exclude={"llm_api_key"}) == LLMEndpointConfig(
+        model="gpt-3.5-turbo-0125",
+        llm_base_url=None,
+        llm_api_key=None,
+        max_input_tokens=2000,
+        max_output_tokens=2000,
+        temperature=0.7,
+        streaming=True,
+    ).model_dump(exclude={"llm_api_key"})
 
 
 def test_default_retrievalconfig():
@@ -23,4 +20,6 @@ def test_default_retrievalconfig():
 
     assert config.max_files == 20
     assert config.prompt is None
-    assert config.llm_config == LLMEndpointConfig()
+    assert config.llm_config.model_dump(
+        exclude={"llm_api_key"}
+    ) == LLMEndpointConfig().model_dump(exclude={"llm_api_key"})
